fix(dashboard): 404 the static aave stakeholders route when AAVE is disabled
